+++ apps/docs/src/components/mdx/mdx-components/browser-mockup.tsx
@@ -253,76 +253,3 @@ export const DemoCard = memo(
 );
 
 DemoCard.displayName = "DemoCard";
-
-// Default Demo Layout Component
-export interface DemoLayoutProps {
-  children: React.ReactNode;
-  logo?: string;
-  title?: string;
-  headerActions?: React.ReactNode;
-}
-
-export const DemoLayout = memo(({ 
-  children, 
-  logo = "⚡", 
-  title = "SSGOI Demo",
-  headerActions 
-}: DemoLayoutProps) => {
-  const context = React.useContext(BrowserContext);
-  const isMobile = useMobile();
-  
-  if (!context) return <>{children}</>;
-  
-  const { currentPath, navigate, routes } = context;
-
-  return (
-    <div className="flex flex-col h-full">
-      {/* Header */}
-      <header className="bg-gray-800 border-b border-gray-700">
-        <div className={cn(
-          "mx-auto",
-          isMobile ? "px-3" : "max-w-6xl px-4"
-        )}>
-          <div className="flex items-center justify-between h-14">
-            <div className="flex items-center gap-4">
-              <h1 className={cn(
-                "font-bold text-white flex items-center gap-2",
-                isMobile ? "text-base" : "text-xl"
-              )}>
-                <span>{logo}</span>
-                <span>{title}</span>
-              </h1>
-              <nav className="flex items-center gap-1">
-                {routes.map((route) => (
-                  <button
-                    key={route.path}
-                    onClick={() => navigate(route.path)}
-                    className={cn(
-                      "rounded-md font-medium transition-all",
-                      isMobile ? "px-2 py-1 text-xs" : "px-4 py-2 text-sm",
-                      currentPath === route.path
-                        ? "bg-gray-700 text-white"
-                        : "text-gray-300 hover:bg-gray-700 hover:text-white"
-                    )}
-                  >
-                    {route.label}
-                  </button>
-                ))}
-              </nav>
-            </div>
-            {!isMobile && headerActions && (
-              <div className="flex items-center gap-4">
-                {headerActions}
-              </div>
-            )}
-          </div>
-        </div>
-      </header>
-
-      {/* Main Content */}
-      <main className="flex-1 overflow-auto relative z-0">{children}</main>
-    </div>
-  );
-});
-
-DemoLayout.displayName = "DemoLayout";
